fixed doctype error that occurred from accidentally deleting proxy and moved files
