+++ reactApp/package.json
@@ -50,5 +50,6 @@
     "@storybook/preset-create-react-app": "^4.0.0",
     "@storybook/react": "^6.4.18",
     "html-webpack-plugin": "^5.5.0"
-  }
+  },
+  "proxy":"http://localhost:8080"
 }
