+++ reactApp/src/api/movie-api.js
@@ -1,64 +1,64 @@
 export const login = (username, password) => {
-    return fetch('/api/users', {
-        headers: {
-            'Content-Type': 'application/json'
-        },
-        method: 'post',
-        body: JSON.stringify({ username: username, password: password })
-    }).then(res => res.json())
+  return fetch('/api/users', {
+      headers: {
+          'Content-Type': 'application/json'
+      },
+      method: 'post',
+      body: JSON.stringify({ username: username, password: password })
+  }).then(res => res.json())
 };
 
 export const signup = (username, password) => {
-    return fetch('/api/users?action=register', {
-        headers: {
-            'Content-Type': 'application/json'
-        },
-        method: 'post',
-        body: JSON.stringify({ username: username, password: password })
-    }).then(res => res.json())
+  return fetch('/api/users?action=register', {
+      headers: {
+          'Content-Type': 'application/json'
+      },
+      method: 'post',
+      body: JSON.stringify({ username: username, password: password })
+  }).then(res => res.json())
 };
 
 export const getMovies = () => {
-    return fetch(
-       '/api/movies',{headers: {
-         'Authorization': window.localStorage.getItem('token')
-      }
+  return fetch(
+     '/api/movies',{headers: {
+       'Authorization': window.localStorage.getItem('token')
     }
-    ).then(res => res.json());
-  };
+  }
+  ).then(res => res.json());
+};
 
 export const getUpcomingMovies = () => {
-    return fetch(
-      `/api/upcomingMovies`,{headers: {
-       'Authorization': window.localStorage.getItem('token')
-    }
-  }
-  ).then(res => res.json());
-};
-
-export const getTrendingMovies = () => {
-    return fetch(
-      `/api/trendingMovies`,{headers: {
-       'Authorization': window.localStorage.getItem('token')
-    }
-  }
-  ).then(res => res.json());
-};
-
-export const getTopRatedMovies = () => {
   return fetch(
-    `/api/topRatedMovies`,{headers: {
+    `/api/upcomingMovies`,{headers: {
      'Authorization': window.localStorage.getItem('token')
   }
 }
 ).then(res => res.json());
 };
 
-export const getGenres = () => {
-    return fetch(
-      `/api/genres`,{headers: {
-       'Authorization': window.localStorage.getItem('token')
-    }
+export const getTrendingMovies = () => {
+  return fetch(
+    `/api/trendingMovies`,{headers: {
+     'Authorization': window.localStorage.getItem('token')
   }
-  ).then(res => res.json());
+}
+).then(res => res.json());
+};
+
+export const getTopRatedMovies = () => {
+return fetch(
+  `/api/topRatedMovies`,{headers: {
+   'Authorization': window.localStorage.getItem('token')
+}
+}
+).then(res => res.json());
+};
+
+export const getGenres = () => {
+  return fetch(
+    `/api/genres`,{headers: {
+     'Authorization': window.localStorage.getItem('token')
+  }
+}
+).then(res => res.json());
 };
