+++ reactApp/src/contexts/authContext.js
@@ -1,5 +1,5 @@
 import React, { useState, createContext } from "react";
-import { login, signup } from "./api/movie-api";
+import { login, signup } from "../api/movie-api";
 
 export const AuthContext = createContext(null);
