+++ reactApp/src/pages/favouriteMoviesPage.js
@@ -1,11 +1,11 @@
 import React, { useContext } from "react";
-import PageTemplate from "./components/templateMovieListPage";
-import { MoviesContext } from "./moviesContext";
+import PageTemplate from "../components/templateMovieListPage";
+import { MoviesContext } from "../moviesContext";
 import { useQueries } from "react-query";
-import { getMovie } from "./api/tmdb-api";
-import Spinner from './components/spinner';
-import RemoveFromFavourites from "./components/cardIcons/removeFromFavourites";
-import WriteReview from "./components/cardIcons/writeReview";
+import { getMovie } from "../api/tmdb-api";
+import Spinner from '../components/spinner';
+import RemoveFromFavourites from "../components/cardIcons/removeFromFavourites";
+import WriteReview from "../components/cardIcons/writeReview";
 
 const FavouriteMoviesPage = () => {
   const {favourites: movieIds } = useContext(MoviesContext);
