+++ reactApp/src/pages/homePage.js
@@ -1,9 +1,9 @@
 import React from "react";
-import { getMovies } from "./api/movie-api";
-import PageTemplate from './components/templateMovieListPage';
+import { getMovies } from "../api/movie-api";
+import PageTemplate from '../components/templateMovieListPage';
 import { useQuery } from 'react-query';
-import Spinner from './components/spinner';
-import AddToFavouritesIcon from './components/cardIcons/addToFavourites';
+import Spinner from '../components/spinner';
+import AddToFavouritesIcon from '../components/cardIcons/addToFavourites';
 
 const HomePage = (props) => {
