+++ reactApp/src/pages/movieDetailsPage.js
@@ -1,11 +1,11 @@
 import React from "react";
 import { useParams } from 'react-router-dom';
-import MovieDetails from "./components/movieDetails";
-import PageTemplate from "./components/templateMoviePage";
-import useMovie from "./hooks/useMovie";
-import { getMovie } from './api/tmdb-api';
+import MovieDetails from "../components/movieDetails";
+import PageTemplate from "../components/templateMoviePage";
+import useMovie from "../hooks/useMovie";
+import { getMovie } from '../api/tmdb-api';
 import { useQuery } from "react-query";
-import Spinner from './components/spinner';
+import Spinner from '../components/spinner';
 import { Link } from 'react-router-dom';
 
 const MovieDetailsPage = (props) => {
