+++ reactApp/src/pages/signUpPage.js
@@ -1,6 +1,6 @@
 import React, { useContext, useState } from "react";
 import { Navigate } from "react-router-dom";
-import { AuthContext } from './authContext';
+import { AuthContext } from '../contexts/authContext';
 
 const SignUpPage = props => {
   const context = useContext(AuthContext)
