+++ reactApp/src/privateRoute.js
@@ -1,6 +1,6 @@
 import React, { useContext} from "react";
 import { Route, Navigate } from "react-router-dom";
-import {AuthContext} from './authContext'
+import {AuthContext} from './contexts/authContext'
 
 const PrivateRoute = props => {
   const context = useContext(AuthContext)
